+++ Sources/HummingbirdFoundation/Cookies/Cookie.swift
@@ -1,15 +1,5 @@
 import Foundation
 
-extension HBDateCache {
-    static var rfc1123Formatter: DateFormatter = {
-         let formatter = DateFormatter()
-         formatter.locale = Locale(identifier: "en_US_POSIX")
-         formatter.dateFormat = "EEE, d MMM yyy HH:mm:ss z"
-         formatter.timeZone = TimeZone(secondsFromGMT: 0)
-         return formatter
-     }()
-}
-
 /// Structure holding a single cookie
 public struct HBCookie: CustomStringConvertible {
     public enum SameSite: String {
