+++ Sources/HummingbirdFoundation/Utils/DateCache+Date.swift
@@ -1,0 +1,11 @@
+import Foundation
+
+extension HBDateCache {
+    static var rfc1123Formatter: DateFormatter = {
+        let formatter = DateFormatter()
+        formatter.locale = Locale(identifier: "en_US_POSIX")
+        formatter.dateFormat = "EEE, d MMM yyy HH:mm:ss z"
+        formatter.timeZone = TimeZone(secondsFromGMT: 0)
+        return formatter
+    }()
+}
